+++ package.json
@@ -7,7 +7,7 @@
         "compile": "babel src --presets react,es2015 --out-dir static",
         "watch": "babel src --presets react,es2015 --out-dir static --watch",
         "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "nodemon -w server.js server.js"
+        "start": "nodemon -w server server/server.js"
     },
     "author": "",
     "license": "ISC",
