+++ server/server.js
@@ -1,6 +1,7 @@
 const express = require("express");
 
 const MongoClient = require("mongodb").MongoClient;
+const Issue = require("./issue.js");
 
 const app = express();
 
@@ -8,72 +9,6 @@ const bodyParser = require("body-parser");
 
 app.use(express.static("static"));
 app.use(bodyParser.json());
-
-const issues = [
-  {
-    id: 1,
-    status: "Open",
-    owner: "Ravan",
-    created: new Date("2016-08-15"),
-    effort: 5,
-    completionDate: undefined,
-    title: "Error in console when clicking Add"
-  },
-  {
-    id: 2,
-    status: "Assigned",
-    owner: "Eddie",
-    created: new Date("2016-08-16"),
-    effort: 14,
-    completionDate: new Date("2016-08-30"),
-    title: "Missing bottom border on panel"
-  }
-];
-
-const validIssueStatus = {
-  New: true,
-  Open: true,
-  Assigned: true,
-  Fixed: true,
-  Verified: true,
-  Clsed: true
-};
-
-const issueFieldType = {
-  id: "required",
-  status: "required",
-  owner: "required",
-  effort: "optional",
-  created: "required",
-  completionDate: "optional",
-  title: "required"
-};
-
-const keyValue = {
-  owner: "Propietario",
-  title: "Título"
-};
-
-function validateIssue(issue) {
-  for (const field in issueFieldType) {
-    const type = issueFieldType[field];
-    console.log(
-      "Campo, Tipo, Valor ==>> %s, %s, %s",
-      field,
-      type,
-      issue[field]
-    );
-    if (!type) {
-      delete issue[field];
-    } else if (type === "required" && !issue[field]) {
-      return `${keyValue[field]} es requerido.`;
-    }
-  }
-  if (!validIssueStatus[issue.status]) {
-    return `${issue.status} no es un valido estado.`;
-  }
-  return null;
-}
 
 app.get("/api/issues", (req, res) => {
   db.collection("issues")
@@ -98,22 +33,33 @@ app.get("/api/issues", (req, res) => {
 app.post("/api/issues", (req, res) => {
   const newIssue = req.body;
   console.log("##=> newIssue: %O", newIssue);
-
-  newIssue.id = issues.length + 1;
   newIssue.created = new Date();
   if (!newIssue.status) {
     newIssue.status = "New";
   }
 
-  const err = validateIssue(newIssue);
+  const err = Issue.validateIssue(newIssue);
   if (err) {
     res.status(442).json({ message: `Petición invalida: ${err}` });
     return;
   }
 
-  issues.push(newIssue);
-
-  res.json(newIssue);
+  db.collection("issues")
+    .insertOne(newIssue)
+    .then((result) =>
+      db
+        .collection("issues")
+        .find({ _id: result.insertedId })
+        .limit(1)
+        .next()
+    )
+    .then((newIssue) => {
+      res.json(newIssue);
+    })
+    .catch((error) => {
+      console.log(error);
+      res.status(500).json({ message: `Internal Server Error: ${error}` });
+    });
 });
 
 let db;
